Fix ccd-definitions package
