+++ ccd-definitions/src/main/java/uk/gov/hmcts/reform/ccd/DevelopmentCcdConfig.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.ccd.ccd;
+package uk.gov.hmcts.reform.ccd;
 
 import uk.gov.hmcts.ccd.sdk.types.CCDConfig;
 import uk.gov.hmcts.ccd.sdk.types.ConfigBuilder;
 import uk.gov.hmcts.ccd.sdk.types.Webhook;
-import uk.gov.hmcts.reform.ccd.ccd.model.CaseData;
-import uk.gov.hmcts.reform.ccd.ccd.model.State;
-import uk.gov.hmcts.reform.ccd.ccd.model.UserRole;
+import uk.gov.hmcts.reform.ccd.model.CaseData;
+import uk.gov.hmcts.reform.ccd.model.State;
+import uk.gov.hmcts.reform.ccd.model.UserRole;
 
 public class DevelopmentCcdConfig implements CCDConfig<CaseData, State, UserRole> {
