+++ ccd-definitions/src/main/java/uk/gov/hmcts/reform/ccd/model/CaseData.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.ccd.ccd.model;
+package uk.gov.hmcts.reform.ccd.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
