+++ ccd-definitions/src/main/java/uk/gov/hmcts/reform/ccd/model/UserRole.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.ccd.ccd.model;
+package uk.gov.hmcts.reform.ccd.model;
 
 import uk.gov.hmcts.ccd.sdk.types.HasRole;
